Implementing RT Core-based ray tracing rendering, Step 19

+++ src/vulkan_manager/buffer_manager.cpp
@@ -350,7 +350,7 @@ void RandomBufferManager::init()
 {
 	std::vector<float> random_table(this->size);
 #pragma omp parallel for
-	for (size_t i = 0; i < this->size; i++)
+	for (int i = 0; i < this->size; i++)
 	{
 		random_table[i] = getRandomNumber(0.0, 1.0);
 	}
